Initialize console if password is required (#121)
In absence of console or when console reference line is set
to nil or noPwd is set to true, password prompt is disabled
and when console is not initialized, password cannot be read
from user. This is why when -i is specified, console is not
initialized and we do not see password prompt.
This commit initializes console even if password is required
so that password can be read interactively from the user.