Set only available info in env vars
Earlier commit was setting env vars only if endpoint and user info
was associated with context. Context however may or may not have
user associated with it if end user wants to use trusted authentication.
In this commit, we are setting only the available info and inform user
of context being used